refactor: split python ast analysis internals
